fix: toast position (#1288)
We've encountered a bug where, upon adding a new toast, all existing
toasts are overwritten with the same fields, including the key. This
results in an array where all toasts appear identical.
<img width="851" alt="old"
src="https://github.com/user-attachments/assets/475aabc7-e9c2-4c4f-a855-e240f043f38d"
/>

i think this is a bug with signals when we try to set of destructurred
array
I believe the issue arises from the use of signals when we attempt to
set a destructured array:
`$toasts.set([{
    key: toastId,
    show: true,
    topPosition: 0,
    alert,
    icon,
    type,
    text,
    title,
    actionLabel,
    onAction,
    onClose: onRemove,
    onLayout
  }, ...$toasts.get() || []]
`
To resolve this, I using unshift() to add the new toast to the beginning
of the array and then setting the updated array. This way, we avoid the
duplication issue.

After implementing this fix, the result is as expected:

<img width="818" alt="new"
src="https://github.com/user-attachments/assets/b27101b8-278d-48e8-aad8-3e87bb01804d"
/>



https://github.com/user-attachments/assets/067a2ea4-13c4-489a-930c-0785077d36f2